+++ examples/pup/hub_essentialhub/imu_read_vector.py
@@ -1,0 +1,14 @@
+from pybricks.hubs import EssentialHub
+from pybricks.tools import wait
+
+# Initialize the hub.
+hub = EssentialHub()
+
+# Get the acceleration vector.
+print(hub.imu.acceleration())
+
+# Get the angular velocity vector.
+print(hub.imu.angular_velocity())
+
+# Wait so we can see what we printed
+wait(5000)