+++ examples/pup/hub_essentialhub/imu_tilt_blast.py
@@ -1,12 +1,11 @@
-# ExampleHub = TechnicHub PrimeHub EssentialHub
-from pybricks.hubs import ExampleHub
+from pybricks.hubs import EssentialHub
 from pybricks.tools import wait
 from pybricks.geometry import Axis
 
 # Initialize the hub. In this case, specify that the hub is mounted with the
 # top side facing forward and the front side facing to the right.
 # For example, this is how the hub is mounted in BLAST in the 51515 set.
-hub = ExampleHub(top_side=Axis.X, front_side=-Axis.Y)
+hub = EssentialHub(top_side=Axis.X, front_side=-Axis.Y)
 
 while True:
     # Read the tilt values. Now, the values are 0 when BLAST stands upright.