+++ examples/pup/hub_essentialhub/light_blink.py
@@ -1,10 +1,9 @@
-# ExampleHub = MoveHub CityHub TechnicHub PrimeHub EssentialHub
-from pybricks.hubs import ExampleHub
+from pybricks.hubs import EssentialHub
 from pybricks.parameters import Color
 from pybricks.tools import wait
 
 # Initialize the hub
-hub = ExampleHub()
+hub = EssentialHub()
 
 # Keep blinking red on and off.
 hub.light.blink(Color.RED, [500, 500])